test(post_processing): split output test module
Break the large output test file into focused modules so related cases stay easier to navigate and maintain without changing behavior.
